split userrepo into user and role repos
