+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/db/MappingHelpers.kt
@@ -8,3 +8,4 @@ inline fun <reified T : Enum<T>> parseEnum(name: String): T
             .firstOrNull { name.equals(it.name, ignoreCase = true) }
             ?: throw UnknownEnumValue(name, T::class.simpleName ?: "[unknown]")
 }
+
